+++ app/src/main/res/layout/activity_destinations.xml
@@ -11,29 +11,43 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:theme="@style/AppTheme.AppBarOverlay">
+        <android.support.design.widget.CollapsingToolbarLayout
+            android:id="@+id/collapsingbar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:theme="@style/ThemeOverlay.AppCompat.Dark"
+            app:contentScrim="?attr/colorPrimary"
+            app:layout_scrollFlags="scroll|exitUntilCollapsed"
+            android:fitsSystemWindows="true">
+
+        <FrameLayout
+            android:id="@+id/photo_container"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:background="@color/photo_placeholder"
+            app:layout_collapseMode="parallax"
+            android:fitsSystemWindows="true">
+            <ImageView
+                android:id="@+id/destination_photo"
+                android:layout_width="match_parent"
+                android:layout_height="@dimen/detail_photo_height"
+                android:adjustViewBounds="true"
+                android:src="@drawable/sf"
+                android:scaleType="centerCrop"
+                android:fitsSystemWindows="true"
+                app:layout_collapseMode="parallax"/>
+        </FrameLayout>
 
         <android.support.v7.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
             android:background="?attr/colorPrimary"
+            app:layout_collapseMode="pin"
             app:popupTheme="@style/AppTheme.PopupOverlay" />
+        </android.support.design.widget.CollapsingToolbarLayout>
 
     </android.support.design.widget.AppBarLayout>
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Try these destinations.."
-        android:id="@+id/destination_text"
-        android:gravity="center_vertical"
-        android:textSize="24sp"
-        android:textColor="@color/colorPrimary"
-        android:textStyle="bold"
-        android:padding="@dimen/activity_horizontal_margin"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
-        />
 
     <include layout="@layout/content_destinations" />
