+++ app/src/main/res/layout/activity_destinations_bac.xml
@@ -1,0 +1,9 @@
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/fragment"
+    android:name="com.github.gripsack.android.ui.destinations.DestinationsFragment"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+
+    tools:layout="@layout/fragment_destinations_bac" />
