+++ app/src/main/res/layout/activity_drawer.xml
@@ -9,7 +9,7 @@
     tools:openDrawer="start">
 
     <include
-        layout="@layout/activity_fragment"
+        layout="@layout/activity_fragment_bac"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
