+++ app/src/main/res/layout/activity_fragment_bac.xml
@@ -1,0 +1,89 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:fitsSystemWindows="true"
+    >
+
+    <android.support.design.widget.AppBarLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay"
+        android:fitsSystemWindows="true">
+        <android.support.design.widget.CollapsingToolbarLayout
+            android:id="@+id/collapsingbar"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:theme="@style/ThemeOverlay.AppCompat.Dark"
+            app:contentScrim="?attr/colorPrimary"
+            app:layout_scrollFlags="scroll|exitUntilCollapsed"
+            android:fitsSystemWindows="true">
+
+            <FrameLayout
+                android:id="@+id/photo_container"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:background="@color/photo_placeholder"
+                app:layout_collapseMode="parallax"
+                android:fitsSystemWindows="true">
+                <ImageView
+                    android:id="@+id/photoCollapse"
+                    android:layout_width="match_parent"
+                    android:layout_height="@dimen/detail_photo_height"
+                    android:adjustViewBounds="true"
+                    android:scaleType="centerCrop"
+                    android:src="@drawable/travel_photo"
+                    android:fitsSystemWindows="true"
+                    app:layout_collapseMode="parallax"/>
+                <TextView
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:text="Try these destinations..!"
+                    android:id="@+id/text"
+                    android:layout_gravity="bottom"
+                    android:gravity="center_horizontal"
+                    android:layout_marginBottom="16dp"
+                    android:textSize="26sp"
+                    android:textColor="@color/color_white"
+                    android:textStyle="bold"
+                    android:padding="@dimen/activity_horizontal_margin"
+                    />
+            </FrameLayout>
+
+            <android.support.v7.widget.Toolbar
+                android:id="@+id/toolbar"
+                android:layout_width="match_parent"
+                android:layout_height="?attr/actionBarSize"
+                app:layout_collapseMode="pin"
+                app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
+                app:layout_scrollFlags="scroll|enterAlways">
+            <EditText
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:id="@+id/toolbarText"
+                android:background="@color/color_white"
+                android:outlineProvider="bounds"
+                android:hint="Where'd you like to go?"
+                android:textSize="12sp"
+                android:padding="10dp"
+                android:textColorHint="#a7a7af"
+                android:clickable="true"
+                />
+            </android.support.v7.widget.Toolbar>
+
+        </android.support.design.widget.CollapsingToolbarLayout>
+
+    </android.support.design.widget.AppBarLayout>
+
+
+    <include
+        android:id="@+id/fragment_container"
+        layout="@layout/fragment_container"/>
+
+    <include layout="@layout/fab"/>
+
+
+
+</android.support.design.widget.CoordinatorLayout>
