+++ app/src/main/res/layout/destination_item.xml
@@ -6,7 +6,7 @@
 
     <ImageView
         android:layout_width="match_parent"
-        android:layout_height="180dp"
+        android:layout_height="200dp"
         android:id="@+id/destination_image"
         android:adjustViewBounds="true"
         android:scaleType="centerCrop"
@@ -21,7 +21,7 @@
         android:gravity="center"
         android:id="@+id/destination_name"
         android:textSize="22sp"
-        android:textColor="#FFFF"
+        android:textColor="@color/color_white"
         android:textStyle="bold"/>
 
 </FrameLayout>
