+++ app/src/main/res/layout/explore_grid_item.xml
@@ -1,66 +1,47 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/card_view"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:cardUseCompatPadding="true"
-    android:clickable="true"
-    android:focusable="true"
-    app:cardBackgroundColor="@android:color/white"
-    app:cardElevation="4dp"
-    app:cardCornerRadius="3dp"
-    android:foreground="?android:attr/selectableItemBackground">
-    <RelativeLayout
+
+
+    <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:background="#F3E5F5">
         <ImageView
             android:layout_width="match_parent"
-            android:layout_height="90dp"
-            android:id="@+id/item_image"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentLeft="true"
+            android:layout_height="200dp"
             android:scaleType="centerCrop"
-            />
+            android:id="@+id/item_image"
+            android:background="@color/photo_placeholder"
+            android:adjustViewBounds="true"
+          />
         <TextView
+            android:id="@+id/item_name"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:id="@+id/item_name"
-            android:gravity="center_horizontal"
-            android:layout_below="@id/item_image"
-            android:minLines="2"/>
-        <ImageButton
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/item_bucketlist"
-            android:layout_below="@id/item_name"
-            android:src="@drawable/ic_action_bucket"/>
+            android:padding="@dimen/keyline_margin"
+            android:ellipsize="end"
+            android:maxLines="4"
+            android:textAppearance ="@style/TextAppearance.AppCompat.Title"
+            android:textColor="@color/color_white"
+            android:textStyle="bold"
+            android:fontFamily="sans-serif"
+            android:gravity="left|start"
+            android:layout_gravity="bottom"
+            android:paddingEnd="@dimen/keyline_margin"
+            android:paddingStart="@dimen/keyline_margin" />
         <ImageButton
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:id="@+id/item_like"
-            android:layout_toRightOf="@id/item_bucketlist"
-            android:layout_below="@id/item_name"
+           android:layout_gravity="end"
+            android:background="@android:color/transparent"
             android:src="@drawable/ic_action_like"/>
         <ImageButton
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:id="@+id/item_add"
-            android:layout_toRightOf="@id/item_like"
+            android:id="@+id/item_bucketlist"
+            android:background="@android:color/transparent"
             android:layout_below="@id/item_name"
-            android:src="@drawable/ic_action_add"/>
+            android:src="@drawable/ic_action_bucket"/>
 
-        <ImageButton
-            android:id="@+id/btnDelete"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@android:drawable/ic_menu_close_clear_cancel"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentRight="true"
-            android:background="@null"
-            android:layout_margin="10dp"
-            android:visibility="gone"/>
+    </FrameLayout>
 
-    </RelativeLayout>
-</android.support.v7.widget.CardView>
