+++ app/src/main/res/layout/fragment_explore.xml
@@ -7,11 +7,11 @@
     tools:context="com.github.gripsack.android.ui.destinations.DestinationsFragment"
     tools:showIn="@layout/activity_destinations">
 
+
     <android.support.v7.widget.RecyclerView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true"
+        android:layout_below="@id/text"
         android:id="@+id/destination_recycler"/>
