+++ app/src/main/res/menu/menu_explore.xml
@@ -5,5 +5,8 @@
         android:id="@+id/action_search"
         android:icon="@android:drawable/ic_menu_search"
         android:title="Search"
+        android:orderInCategory="100"
         app:showAsAction="ifRoom|collapseActionView" />
+
+
 </menu>
