+++ app/src/main/res/values/colors.xml
@@ -8,4 +8,8 @@
     <color name="grey_100">#F5F5F5</color>
     <color name="grey_300">#E0E0E0</color>
 
+    <color name="photo_placeholder">#ccc</color>
+
+    <color name="color_white">#FFF</color>
+
 </resources>
